Separate some utils method into Cornerstone_Utils
